+++ lib/plist_read_write/version.rb
@@ -1,3 +1,3 @@
 module PListReadWrite
-  VERSION = "0.0.1"
+  VERSION = "0.0.1.1"
 end
